Do not fail the install_puppet.sh on 2th run
The pip-python symbolic link creation is the last
step in the install_puppet.sh on Fedora.
If the symbolic link already exists,
the exit status of the script will be failed.

After this change the pip-python link will be created only,
when it does not exists.

+++ install_puppet.sh
@@ -73,7 +73,7 @@ function setup_puppet_fedora {
     # work-around is to just symlink pip-python to "fool" it.
     # See upstream issue:
     #  https://tickets.puppetlabs.com/browse/PUP-1082
-    ln -s /usr/bin/pip /usr/bin/pip-python
+    [ -e /usr/bin/pip-python ] || ln -s /usr/bin/pip /usr/bin/pip-python
 }
 
 function setup_puppet_rhel6 {
